% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_api.R
\name{query_api}
\alias{query_api}
\title{Query API}
\usage{
query_api(
  url,
  username,
  password,
  request_type = "GET",
  timeout_seconds = 300,
  headers = "application/octet-stream",
  output_dir = NULL
)
}
\arguments{
\item{url}{A character vector containing the URL to query the MM API.}

\item{username}{A character vector containing the MM API username.}

\item{password}{A character vector containing the MM API password.}

\item{request_type}{A character vector containing the request type - either
"get" (Default) or "post".}

\item{timeout_seconds}{A number giving the time in seconds, when the request
results in a timeout. The default value is 300.}

\item{headers}{A character vector with one element. The default value is
"application/octet-stream".}

\item{output_dir}{A character vector containing the output directory and the
file name. It is used for saving images as .png and .geoTiff
files as well as for saving netcdf files. The default
value is NULL.}
}
\value{
A list containing the API response.
}
\description{
Query weather data from the Meteomatics API. The full URL including the
Meteomatics username and password is required as input
}
\examples{
\dontrun{
username <- 'r-community'
password <- 'Utotugode673'
datetime <- strftime(ISOdatetime(year=as.integer(strftime(lubridate::today(),'\%Y')),
                                 month=as.integer(strftime(lubridate::today(),'\%m')),
                                 day=as.integer(strftime(lubridate::today(),'\%d')),
                                 hour = 00, min = 00, sec = 00, tz = 'UTC'),
                     format = '\%Y-\%m-\%dT\%H:\%M:\%OSZ', tz = 'UTC')
url <- paste0(sprintf('https://\%s:\%s@api.meteomatics.com/', username, password),
                      datetime, '/t_2m:C/52.520551,13.461804/csv')
query_api(url, username, password)
}
}
