% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_available_time_ranges.R
\name{query_available_time_ranges}
\alias{query_available_time_ranges}
\title{Query the Available Time Ranges}
\usage{
query_available_time_ranges(parameters, username, password, model)
}
\arguments{
\item{parameters}{A list of strings containing the parameters of interest
list("t_2m:C", "dew_point_2m:C", "relative_humidity_1000hPa:p",
"precip_1h:mm").}

\item{username}{A character vector containing the MM API username.}

\item{password}{A character vector containing the MM API password.}

\item{model}{A character vector containing the model of interest. The default
value is NULL, meaning that the model mix is selected.}
}
\value{
A list of available time ranges.
}
\description{
Retrieve the available time ranges from the Meteomatics Weather API
}
\examples{
\dontrun{
username <- "r-community"
password <- "Utotugode673"
parameters <- list("t_2m:C", "dew_point_2m:C", "relative_humidity_2m:p", "precip_1h:mm",
                "global_rad:W", "evapotranspiration_1h:mm")
model <- "ukmo-euro4"
available_time_ranges <- query_available_time_ranges(parameters, username, password, model)
}
}
