% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_init_date.R
\name{query_init_date}
\alias{query_init_date}
\title{Query the Latest Initial Times}
\usage{
query_init_date(
  startdate,
  enddate,
  interval,
  parameters,
  username,
  password,
  model
)
}
\arguments{
\item{startdate}{An ISOdatetime of the format \%Y-\%m-\%dT\%H:\%M:\%OSZ. The start
date gets converted into UTC if another timezone is selected.}

\item{enddate}{An ISOdatetime of the format \%Y-\%m-\%dT\%H:\%M:\%OSZ. The end date
gets converted into UTC if another timezone is selected.}

\item{interval}{A string e.g. of the form "PT1H" specifying the
interval of the parameter.}

\item{parameters}{A list of strings containing the parameters of interest.}

\item{username}{A string containing the MM API username.}

\item{password}{A string containing the MM API password.}

\item{model}{A string containing the model of interest. The default
value is NULL, meaning that the model mix is selected.}
}
\value{
A list of initial dates.
}
\description{
Retrieve a list from the Meteomatics Weather API of the latest initial times
}
\examples{
\dontrun{
username <- "r-community"
password <- "Utotugode673"
time_zone <- "UTC"
startdate <- ISOdatetime(year = as.integer(strftime(lubridate::today(), '\%Y')),
                         month = as.integer(strftime(lubridate::today(), '\%m')),
                         day = as.integer(strftime(lubridate::today(), '\%d')) - 1,
                         hour = 00, min = 00, sec = 00, tz = time_zone)
enddate <- ISOdatetime(year = as.integer(strftime(lubridate::today(), '\%Y')),
                       month = as.integer(strftime(lubridate::today(), '\%m')),
                       day = as.integer(strftime(lubridate::today(), '\%d')),
                       hour = 00, min = 00, sec = 00, tz = time_zone)
interval <- "PT3H"
parameters <- list("t_2m:C", "dew_point_2m:C", "relative_humidity_2m:p", "precip_1h:mm")
model <- "ecmwf-ens"
init_dates <- query_init_date(startdate, enddate, interval, parameters, username,
                              password, model)
head(init_dates)
}
}
