% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_lightnings.R
\name{query_lightnings}
\alias{query_lightnings}
\title{Query a List of Lightnings}
\usage{
query_lightnings(
  startdate,
  enddate,
  lat_N,
  lon_W,
  lat_S,
  lon_E,
  username,
  password,
  request_type = "GET"
)
}
\arguments{
\item{startdate}{An ISOdatetime of the format \%Y-\%m-\%dT\%H:\%M:\%OSZ. The start
date gets converted into UTC if another timezone is selected.}

\item{enddate}{An ISOdatetime of the format \%Y-\%m-\%dT\%H:\%M:\%OSZ. The end date
gets converted into UTC if another timezone is selected.}

\item{lat_N}{A number giving the most northern coordinate of the desired grid.}

\item{lon_W}{A number giving the most western coordinate of the desired grid.}

\item{lat_S}{A number giving the most southern coordinate of the desired grid.}

\item{lon_E}{A number giving the most eastern coordinate of the desired grid.}

\item{username}{A character vector containing the MM API username.}

\item{password}{A character vector containing the MM API password.}

\item{request_type}{A character vector containing the request type - either
"get" (default) or "post".}
}
\value{
A DataFrame containing the valid date, coordinates and current in kA
of each lightning strike.
}
\description{
Retrieve a list of lightning strikes from the Meteomatics Weather API
}
\examples{
\dontrun{
username <- "r-community"
password <- "Utotugode673"
startdate <- ISOdatetime(year = as.integer(strftime(lubridate::today(), '\%Y')),
                         month = as.integer(strftime(lubridate::today(), '\%m')),
                         day = as.integer(strftime(lubridate::today(), '\%d')) - 1,
                         hour = 00, min = 00, sec = 00, tz = 'UTC')
enddate <- ISOdatetime(year = as.integer(strftime(lubridate::today(), '\%Y')),
                       month = as.integer(strftime(lubridate::today(), '\%m')),
                       day = as.integer(strftime(lubridate::today(), '\%d')),
                       hour = 00, min = 00, sec = 00, tz = 'UTC')
lat_N <- 50
lon_W <- -15
lat_S <- 20
lon_E <- 10
query_lightnings(startdate, enddate, lat_N, lon_W, lat_S, lon_E, username,
                password)
}
}
