% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_netcdf.R
\name{query_netcdf}
\alias{query_netcdf}
\title{Query a Grid Time Series with netcdf as Output Format}
\usage{
query_netcdf(
  filename,
  startdate,
  enddate,
  interval,
  parameter_netcdf,
  lat_N,
  lon_W,
  lat_S,
  lon_E,
  resolution,
  username,
  password,
  model = NULL,
  ens_select = NULL,
  interp_select = NULL,
  request_type = "GET",
  cluster_select = NULL,
  ...
)
}
\arguments{
\item{filename}{A character vector specifying the directory and filename of
the output file. For example, it can be of the form:
paste("/your/desired/directory", "filename.tiff", sep = "/").}

\item{startdate}{An ISOdatetime of the format \%Y-\%m-\%dT\%H:\%M:\%OSZ. The start
date gets converted into UTC if another timezone is selected.}

\item{enddate}{An ISOdatetime of the format \%Y-\%m-\%dT\%H:\%M:\%OSZ. The end date
gets converted into UTC if another timezone is selected.}

\item{interval}{A character vector e.g. of the form "PT1H" specifying the
interval of the parameter.}

\item{parameter_netcdf}{A string containing the parameter of interest.}

\item{lat_N}{A number giving the most northern coordinate of the desired grid.}

\item{lon_W}{A number giving the most western coordinate of the desired grid.}

\item{lat_S}{A number giving the most southern coordinate of the desired grid.}

\item{lon_E}{A number giving the most eastern coordinate of the desired grid.}

\item{resolution}{The resolution in latitude and longitude direction. If the
resolution is in degrees then in the form "0.1,0.5", or if the
resolution is in the number of latitudes/longitudes then in
the form "4x5".}

\item{username}{A character vector containing the MM API username.}

\item{password}{A character vector containing the MM API password.}

\item{model}{A character vector containing the model of interest. The default
value is NULL, meaning that the model mix is selected.}

\item{ens_select}{A character vector containing the ensembles of interest. The
default value is NULL. Possible inputs are for example:
"median"; "member:5"; "member:1-50"; "member:0"; "mean";
"quantile0.2".}

\item{interp_select}{A character vector specifying the interpolation: The
default value is NULL. A possible input is:
"gradient_interpolation".}

\item{request_type}{A character vector containing the request type - either
"get" (Default) or "post".}

\item{cluster_select}{A character vector containing the cluster of interest.
The default value is NULL. Possible inputs are for example:
"cluster:1"; "cluster:1-6".}

\item{...}{A list of additional arguments. One possible argument is
calibrated = TRUE.}
}
\value{
A series of netcdf files on a grid showing values of the specified
parameter.
}
\description{
Retrieve a grid time series from the Meteomatics Weather API and save it as
netcdf files
}
\examples{
\dontrun{
username <- "r-community"
password <- "Utotugode673"
filename <- paste("/Users/Desktop", "grid.tiff", sep = "/")
startdate <- ISOdatetime(year = as.integer(strftime(lubridate::today(), '\%Y')),
                         month = as.integer(strftime(lubridate::today(), '\%m')),
                         day = as.integer(strftime(lubridate::today(), '\%d')) - 1,
                         hour = 00, min = 00, sec = 00, tz = 'UTC')
enddate <- ISOdatetime(year = as.integer(strftime(lubridate::today(), '\%Y')),
                       month = as.integer(strftime(lubridate::today(), '\%m')),
                       day = as.integer(strftime(lubridate::today(), '\%d')),
                       hour = 00, min = 00, sec = 00, tz = 'UTC')
interval <- "PT1H"
parameter <- "evapotranspiration_1h:mm"
lat_N <- 50
lon_W <- -15
lat_S <- 20
lon_E <- 10
resolution <- "4x5" # or "1,0.5"
model <- "mix"
query_netcdf(filename, startdate, enddate, interval, parameter,
             lat_N, lon_W, lat_S, lon_E, resolution, username, password,
             model = model)
}
}
