% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_station_list.R
\name{query_station_list}
\alias{query_station_list}
\title{Query all the available stations for the specified parameters}
\usage{
query_station_list(
  username,
  password,
  source = NULL,
  parameters = NULL,
  startdate = NULL,
  enddate = NULL,
  location = NULL,
  request_type = "GET",
  elevation = NULL,
  id = NULL
)
}
\arguments{
\item{username}{A character vector containing the MM API username.}

\item{password}{A character vector containing the MM API password.}

\item{source}{The source of the data, by default mix-obs. String.}

\item{parameters}{A list of strings containing the parameters of interest:
list("t_2m:C", "dew_point_2m:C", "relative_humidity_1000hPa:p",
"precip_1h:mm"). If not specified, then the default value is NULL.}

\item{startdate}{An ISOdatetime of the format \%Y-\%m-\%dT\%H:\%M:\%OSZ. The start
date gets converted into UTC if another timezone is selected.
If not specified, then the default value is NULL.}

\item{enddate}{An ISOdatetime of the format \%Y-\%m-\%dT\%H:\%M:\%OSZ. The end date
gets converted into UTC if another timezone is selected. If
not specified, then the default value is NULL.}

\item{location}{The location as a string, for example location <- ‘47,8_40,15’,
location <- ‘uk’ or location <- "47.3,9.3",
location <- c("47.3", "9.3").}

\item{request_type}{A character vector containing the request type - either
"get" (Default) or "post".}

\item{elevation}{Height above sea level. It will look first for stations that
are close to this height. Integer or float.}

\item{id}{This is the WMO id, e.g. 66810.}
}
\value{
A DataFrame containing the stations. The best matching stations to
your request appear on top.
}
\description{
Retrieve a station list from the Meteomatics Weather API
}
\examples{

username <- "r-community"
password <- "Utotugode673"
source <- "mix-obs"
#parameter <- list("t_2m:C")
#elevation <- 300
#id <- 66810
location <- "47.3,9.3"
query_station_list(username, password, source = source, location = location)

}
