% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_station_timeseries.R
\name{query_station_timeseries}
\alias{query_station_timeseries}
\title{Query a Timeseries at Specified Stations}
\usage{
query_station_timeseries(
  startdate,
  enddate,
  interval,
  parameters,
  username,
  password,
  model = "mix-obs",
  latlon_tuple_list = NULL,
  wmo_ids = NULL,
  mch_ids = NULL,
  general_ids = NULL,
  hash_ids = NULL,
  metar_ids = NULL,
  amda_ids = NULL,
  cman_ids = NULL,
  temporal_interpolation = NULL,
  spatial_interpolation = NULL,
  on_invalid = NULL,
  request_type = "GET"
)
}
\arguments{
\item{startdate}{An ISOdatetime of the format \%Y-\%m-\%dT\%H:\%M:\%OSZ. The start
date gets converted into UTC if another timezone is selected.}

\item{enddate}{An ISOdatetime of the format \%Y-\%m-\%dT\%H:\%M:\%OSZ. The end date
gets converted into UTC if another timezone is selected.}

\item{interval}{A character vector e.g. of the form "PT1H" specifying the
interval of the parameter.}

\item{parameters}{A list of strings containing the parameters of interest:
list("t_2m:C", "dew_point_2m:C",
"relative_humidity_1000hPa:p", "precip_1h:mm"). If not
specified, then the default value is NULL.}

\item{username}{A character vector containing the MM API username.}

\item{password}{A character vector containing the MM API password.}

\item{model}{A character vector containing the model of interest. The default
value is NULL, meaning that the model mix is selected.}

\item{latlon_tuple_list}{The coordinates, in either of these forms:
"47.3,9.3", c("47.3,9.3", "47.43,9.4") or
list(c(47.3,9.3), c(47.43,9.4)).}

\item{wmo_ids}{WMO Station id as string or list of strings.}

\item{mch_ids}{MCH Station id as string or list of strings.}

\item{general_ids}{General Station id as string or list of strings.}

\item{hash_ids}{HASH Station id as string or list of strings.}

\item{metar_ids}{METAR Station id as string or list of strings.}

\item{amda_ids}{AMDA Station id as string or list of strings.}

\item{cman_ids}{C-MAN Station id as string or list of strings.}

\item{temporal_interpolation}{This parameter specifies the temporal interpolation.
The default value is NULL which means the data is interpolated.
If no interpolation is needed, one can set
temporal_interpolation=none.}

\item{spatial_interpolation}{This parameter specifies the spatial interpolation.
The default value is NULL which means the data is interpolated.
If no interpolation is needed, one can set
spatial_interpolation=none.}

\item{on_invalid}{A character vector specifying the treatment of missing
weather station values. The default value is NULL. If
on_invalid = "fill_with_invalid", missing values are filled
with Na.}

\item{request_type}{A character vector containing the request type - either
"get" (Default) or "post".}
}
\value{
A dataframe containing the valid date and the specified parameters.
}
\description{
Retrieve a time series from the Meteomatics Weather API
}
\examples{
\dontrun{
username <- "r-community"
password <- "Utotugode673"
time_zone <- "UTC"
startdate <- ISOdatetime(year = as.integer(strftime(lubridate::today(), '\%Y')),
                      month = as.integer(strftime(lubridate::today(), '\%m')),
                      day = as.integer(strftime(lubridate::today(), '\%d')) - 1,
                      hour = 00, min = 00, sec = 00, tz = time_zone)
enddate <- ISOdatetime(year = as.integer(strftime(lubridate::today(), '\%Y')),
                     month = as.integer(strftime(lubridate::today(), '\%m')),
                     day = as.integer(strftime(lubridate::today(), '\%d')),
                     hour = 00, min = 00, sec = 00, tz = time_zone)
interval <- "PT1H"
parameters <- list("t_2m:C")
coordinates <- list(c(47.3,9.3), c(47.43,9.4))
wmo_ids <- c("066600", "066810")
mch_ids <- c("LUG", "STG")
metar_ids <- c("LSZH", "LSGG")
amda_ids <- "P100"
cman_ids <- "mrka2"
hash_ids <- c("3817692398")
query_station_timeseries(startdate, enddate, interval, parameters,username,
                         password, model = 'mix-obs',
                         latlon_tuple_list = coordinates, wmo_ids = wmo_ids,
                         mch_ids = mch_ids, hash_ids = hash_ids,
                         metar_ids = metar_ids, amda_ids = amda_ids,
                         cman_ids = cman_ids)
}
}
