% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_time_series.R
\name{query_time_series}
\alias{query_time_series}
\title{Query Time Series}
\usage{
query_time_series(
  coordinate_list,
  startdate,
  enddate,
  interval,
  parameters,
  username,
  password,
  model = NULL,
  ens_select = NULL,
  interp_select = NULL,
  on_invalid = NULL,
  request_type = "GET",
  cluster_select = NULL,
  ...
)
}
\arguments{
\item{coordinate_list}{A character vector or a list of vectors containing the
coordinates/postal codes of the desired location. All
following options separated by a semicolon are examples
of possible input formats: "47.3,9.3";
c("47.3,9.3", "47.43,9.4"); "47.11,11.47+50,10";
"postal_CH9000"; c("postal_CH9000", "postal_CH9014");
c("47.3,9.3", "postal_CH9000"); list(c(47.3,9.3), c(47.43,9.4)).}

\item{startdate}{An ISOdatetime of the format \%Y-\%m-\%dT\%H:\%M:\%OSZ. The start
date gets converted into UTC if another timezone is selected.}

\item{enddate}{An ISOdatetime of the format \%Y-\%m-\%dT\%H:\%M:\%OSZ. The end date
gets converted into UTC if another timezone is selected.}

\item{interval}{A character vector e.g. of the form "PT1H" specifying the
interval of the parameter.}

\item{parameters}{A list of strings containing the parameters of interest
list("t_2m:C", "dew_point_2m:C", "relative_humidity_1000hPa:p",
"precip_1h:mm").}

\item{username}{A character vector containing the MM API username.}

\item{password}{A character vector containing the MM API password.}

\item{model}{A character vector containing the model of interest. The default
value is NULL, meaning that the model mix is selected.}

\item{ens_select}{A character vector containing the ensembles of interest. The
default value is NULL. Possible inputs are for example:
"median"; "member:5"; "member:1-50"; "member:0"; "mean";
"quantile0.2".}

\item{interp_select}{A character vector specifying the interpolation: The
default value is NULL. A possible input is:
"gradient_interpolation".}

\item{on_invalid}{A character vector specifying the treatment of missing
weather station values. The default value is NULL. If
on_invalid = "fill_with_invalid", missing values are filled
with Na.}

\item{request_type}{A character vector containing the request type - either
"get" (Default) or "post".}

\item{cluster_select}{A character vector containing the cluster of interest.
The default value is NULL. Possible inputs are for example:
"cluster:1"; "cluster:1-6".}

\item{...}{A list of additional arguments. One possible argument is
calibrated = TRUE.}
}
\value{
A DataFrame containing the coordinates, validdate and the parameter
values.
}
\description{
Retrieve a time series from the Meteomatics Weather API
}
\examples{
username <- "r-community"
password <- "Utotugode673"
coordinates <- list(c(47.3,9.3), c(47.43,9.4))
startdate <- ISOdatetime(year = as.integer(strftime(lubridate::today(), '\%Y')),
                         month = as.integer(strftime(lubridate::today(), '\%m')),
                         day = as.integer(strftime(lubridate::today(), '\%d')),
                         hour = 00, min = 00, sec = 00, tz = 'UTC')
enddate <- ISOdatetime(year = as.integer(strftime(lubridate::today(), '\%Y')),
                       month = as.integer(strftime(lubridate::today(), '\%m')),
                       day = as.integer(strftime(lubridate::today(), '\%d')) + 1,
                       hour = 00, min = 00, sec = 00, tz = 'UTC')
interval <- "PT1H"
parameters <- list("t_2m:C")
model <- "mix"
query_time_series(coordinates, startdate, enddate, interval, parameters,
                  username, password, model = model, calibrated = TRUE)
}
