% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_user_features.R
\name{query_user_features}
\alias{query_user_features}
\title{Query User Features}
\usage{
query_user_features(username, password)
}
\arguments{
\item{username}{A character vector containing the MM API username.}

\item{password}{A character vector containing the MM API password.}
}
\value{
A named character vector containing the licensing options.
}
\description{
Provides information about your Meteomatics licensing options
}
\examples{
username <- "r-community"
password <- "Utotugode673"
query_user_features(username, password)
}
